Moved sidebar code into its own component
